refactor(client): ファイル名を animated_cavivara.dart にリネーム
AnimatedCavivara クラスに合わせてファイル名を変更し、import を更新した。
